requirements.txt
# Dependências do Projeto Totem Flexmedia
# Sprint 2 - Integração de Sensores e Análise de Dados

# Banco de Dados
psycopg2-binary==2.9.9

# Análise de Dados
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# Visualização e Dashboard
streamlit==1.29.0
plotly==5.18.0

# Utilitários
python-dotenv==1.0.0

